+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## TBD 
+* Fixed logic to handle links that open in new tab for embedded webview.
 * Updated user guide to provide a sample Swift & ObjC code for querying a specific account and return token silently when multiple accounts are present in the cache. 
 * Added client-side fix for the known ADFS PKeyAuth issue. (#1150)