+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## TBD 
+## [1.1.12] - 2020-12-02
 * Added cross-cloud B2B support.
 * Fixed logic to handle links that open in new tab for embedded webview.
 * accountForUsername from MSALPublicClientApplication will return nil back when username is nil or empty